refactor(pty-proxy): apply Rust best-practice review findings
Reviewed against the Rust API Guidelines and clippy idiom:

- stop swallowing stdout write errors: apply_pty returns io::Result and
  the pump exits when the terminal goes away, as it did pre-compositor
- one poisoned-lock policy (recover via PoisonError::into_inner) instead
  of three different silent fallbacks; a poisoned compositor previously
  froze the terminal by discarding all pty output
- erase in one rows_formatted() pass instead of nth() per covered row,
  which re-formatted every preceding row each time
- suggest() partitions top-N before sorting, like search() beside it
- KeyFilter::process returns Cow, dropping a per-keystroke allocation
- PtyProxy::run takes a RunOptions struct instead of three positional
  Options; settings fields sized u32/usize so conversions disappear
- fix most_recent_id's doc having drifted onto most_recent_timestamp
- shared fit-to-width helper for popup and ghost truncation; named
  constants for the Esc disambiguation window, popup minimum width, and
  provider timeouts; drop needless pub(crate) and a no-op test
- trim every comment to 1-2 why-focused lines; rustfmt clean

+++ crates/atuin-client/src/settings.rs
@@ -510,16 +510,15 @@ pub struct PtyProxy {
     #[serde(alias = "enable")]
     pub enabled: bool,
 
-    /// Enable the experimental inline suggestion popup (history-powered
-    /// autocomplete with fish-style ghost text) inside the pty proxy.
-    /// Can also be enabled with `ATUIN_PTY_PROXY_SUGGEST=1`.
+    /// Experimental inline suggestion popup in the pty proxy.
+    /// Also enabled by `ATUIN_PTY_PROXY_SUGGEST=1`.
     pub suggestions: bool,
 
     /// Maximum number of history entries fetched per keystroke.
-    pub suggestions_limit: u64,
+    pub suggestions_limit: u32,
 
     /// Minimum typed characters before suggestions appear.
-    pub suggestions_min_chars: u64,
+    pub suggestions_min_chars: usize,
 }
 
 impl Default for PtyProxy {
